// client/src/enums/CommandEnum.tsx
export enum TurtleCommands {
  FORWARD = 'forward',
  BACKWARD = 'backward',
  LEFT = 'left',
  RIGHT = 'right',
  UP = 'up',
  DOWN = 'down',
  DIG = 'dig',
  DIG_UP = 'dig_up',
  DIG_DOWN = 'dig_down',
  PLACE = 'place',
  PLACE_UP = 'place_up',
  PLACE_DOWN = 'place_down',
  DROP = 'drop',
  DROP_UP = 'drop_up',
  DROP_DOWN = 'drop_down',
  INSPECT = 'inspect',
  INSPECT_UP = 'inspect_up',
  INSPECT_DOWN = 'inspect_down',
  SUCK = 'suck',
  SUCK_UP = 'suck_up',
  SUCK_DOWN = 'suck_down',
}